requirements.txt
Flask
python-dotenv
mysql-connector-python
python-telegram-bot==20.3
gunicorn
